# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Component Repository API",
    "description": "REST API for managing CWL-based workflow components",
    "version": "1.0"
  },
  "paths": {
...
    "/components": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "List Components",
        "operationId": "list_components_components_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComponentKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "items": {
                "enum": [
                  "animal_behavior",
                  "earth_observation",
                  "other",
                  "domain_agnostic"
                ]
              },
              "title": "Domain"
            }
          },
          {
            "name": "favoritesOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Favoritesonly"
            }
          },
          {
            "name": "excludeMine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Excludemine"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "includeParameters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includeparameters"
            }
          },
          {
            "name": "rankAgainst",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rankagainst"
            }
          },
          {
            "name": "favoritesFirst",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Favoritesfirst"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoV1_Annotated_Union_ToolListItemDto__WorkflowListItemDto___FieldInfo_annotation_NoneType__required_True__discriminator__kind____"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/mine": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "List My Components",
        "operationId": "list_my_components_components_mine_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComponentKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "publishedOffset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Publishedoffset"
            }
          },
          {
            "name": "unpublishedOffset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Unpublishedoffset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyItemsResponseDtoV1_Annotated_Union_ToolListItemDto__WorkflowListItemDto___FieldInfo_annotation_NoneType__required_True__discriminator__kind____"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/latest": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "List Latest Components",
        "operationId": "list_latest_components_components_latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComponentKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 6,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ToolListItemDto"
                      },
                      {
                        "$ref": "#/components/schemas/WorkflowListItemDto"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "kind",
                      "mapping": {
                        "tool": "#/components/schemas/ToolListItemDto",
                        "workflow": "#/components/schemas/WorkflowListItemDto"
                      }
                    }
                  },
                  "title": "Response List Latest Components Components Latest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/name-availability": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "Check Name Availability",
        "description": "Whether `name` is still free for a new lineage - of either kind.\n\n`available` reflects the whole catalogue, since names are globally unique across tools\nand workflows. `existing` is only filled in when the caller may actually see that\ncomponent - a name held by somebody else's draft still reads as taken, just without\ndetails.\n\nNOTE: must stay declared above GET /{component_id} - FastAPI matches routes in\ndeclaration order, and \"name-availability\" would otherwise be parsed as a component id.",
        "operationId": "check_name_availability_components_name_availability_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          },
          {
            "name": "excludeId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Excludeid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameAvailabilityDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/components/{component_id}/versions": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "Get Versions",
        "operationId": "get_versions_components__component_id__versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Component Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ToolListItemDto"
                      },
                      {
                        "$ref": "#/components/schemas/WorkflowListItemDto"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "kind",
                      "mapping": {
                        "tool": "#/components/schemas/ToolListItemDto",
                        "workflow": "#/components/schemas/WorkflowListItemDto"
                      }
                    }
                  },
                  "title": "Response Get Versions Components  Component Id  Versions Get"
                }
              }
            }
          },
          "404": {
            "description": "Component not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/components/{component_id}/usages": {
      "get": {
        "tags": [
          "components"
        ],
        "summary": "Get Usages",
        "description": "The workflows running any version of this component - a tool, or a nested\nworkflow - only those the caller may see.",
        "operationId": "get_usages_components__component_id__usages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Component Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComponentUsageDto"
                  },
                  "title": "Response Get Usages Components  Component Id  Usages Get"
                }
              }
            }
          },
          "404": {
            "description": "Component not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentKind": {
        "type": "string",
        "enum": [
          "tool",
          "workflow"
        ],
        "title": "ComponentKind",
        "description": "Which child of the composite a Component is - CWL's Process subclasses, narrowed to\nthe two this repository models. A Tool is a leaf (CommandLineTool / ExpressionTool),\na Workflow is a composite whose steps run other Components."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyItemsResponseDtoV1_Annotated_Union_ToolListItemDto__WorkflowListItemDto___FieldInfo_annotation_NoneType__required_True__discriminator__kind____": {
        "properties": {
          "published": {
            "$ref": "#/components/schemas/PaginatedResponseDtoV1_Annotated_Union_ToolListItemDto__WorkflowListItemDto___FieldInfo_annotation_NoneType__required_True__discriminator__kind____"
          },
          "unpublished": {
            "$ref": "#/components/schemas/PaginatedResponseDtoV1_Annotated_Union_ToolListItemDto__WorkflowListItemDto___FieldInfo_annotation_NoneType__required_True__discriminator__kind____"
          }
        },
        "type": "object",
        "required": [
          "published",
          "unpublished"
        ],
        "title": "MyItemsResponseDtoV1[Annotated[Union[ToolListItemDto, WorkflowListItemDto], FieldInfo(annotation=NoneType, required=True, discriminator='kind')]]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}